# docs/changelog/v1/openapi/agros-api-v1.0.yaml
openapi: 3.1.0
info:
  title: Agri OS Core API
  summary: Deterministic core API for Agri OS
  version: 0.1.0
paths:
...
  /api/v1/shared-resources:
    get:
      tags:
      - SharedResources
      summary: List Shared Resources
      operationId: list_shared_resources_api_v1_shared_resources_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedResourceListResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - SharedResources
      summary: Create Shared Resource
      operationId: create_shared_resource_api_v1_shared_resources_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSharedResourceRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedResourceResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/shared-resources/{shared_resource_id}:
    get:
      tags:
      - SharedResources
      summary: Get Shared Resource
      operationId: get_shared_resource_api_v1_shared_resources__shared_resource_id__get
      parameters:
      - name: shared_resource_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Shared Resource Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedResourceDetail'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - ProjectScopes
      summary: Update Project Scope
      operationId: update_project_scope_api_v1_projects__project_scope_id__patch
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectScopeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectScopeResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/views/customer-360/{customer_id}:
    get:
      tags:
      - Views
      summary: Get Customer 360
      operationId: get_customer_360_api_v1_views_customer_360__customer_id__get
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: string
          title: Customer Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer360View'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/views/available-lots:
    get:
      tags:
      - Views
      summary: Get Available Lots Board
      operationId: get_available_lots_board_api_v1_views_available_lots_get
      parameters:
      - name: skuId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Skuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailableLotListResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/views/pending-fulfillment:
    get:
      tags:
      - Views
      summary: Get Pending Fulfillment
      operationId: get_pending_fulfillment_api_v1_views_pending_fulfillment_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingFulfillmentListResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/views/farm:
    get:
      tags:
      - Views
      summary: Get Farm View
      operationId: get_farm_view_api_v1_views_farm_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmView'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/views/farm-summary-board:
    get:
      tags:
      - Views
      summary: Get Farm Summary Board
      operationId: get_farm_summary_board_api_v1_views_farm_summary_board_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmSummaryBoardResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/views/project-contribution-summary:
    get:
      tags:
      - Views
      summary: Get Project Contribution Summary
      operationId: get_project_contribution_summary_api_v1_views_project_contribution_summary_get
      parameters:
      - name: projectScopeId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectscopeid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectContributionSummaryResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/views/project-pnl-summary:
    get:
      tags:
      - Views
      summary: Get Project Pnl Summary
      operationId: get_project_pnl_summary_api_v1_views_project_pnl_summary_get
      parameters:
      - name: projectScopeId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectscopeid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPnlSummaryResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/views/project-order-allocation-summary:
    get:
      tags:
      - Views
      summary: Get Project Order Allocation Summary
      operationId: get_project_order_allocation_summary_api_v1_views_project_order_allocation_summary_get
      parameters:
      - name: projectScopeId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectscopeid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOrderAllocationSummaryResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/views/project-impacted-actors-summary:
    get:
      tags:
      - Views
      summary: Get Project Impacted Actors Summary
      operationId: get_project_impacted_actors_summary_api_v1_views_project_impacted_actors_summary_get
      parameters:
      - name: projectScopeId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectscopeid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectImpactedActorSummaryResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/views/project-contribution-ledger:
    get:
      tags:
      - Views
      summary: Get Project Contribution Ledger
      operationId: get_project_contribution_ledger_api_v1_views_project_contribution_ledger_get
      parameters:
      - name: projectScopeId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectscopeid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectContributionLedgerResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AvailableLotListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AvailableLotView'
          type: array
          title: Items
      type: object
      required:
      - items
      title: AvailableLotListResponse
...
    Customer360View:
      properties:
        customer:
          $ref: '#/components/schemas/CustomerDetail'
        activePreorders:
          items:
            $ref: '#/components/schemas/PreorderDetail'
          type: array
          title: Activepreorders
        recentOrders:
          items:
            $ref: '#/components/schemas/OrderDetail'
          type: array
          title: Recentorders
        preferences:
          items:
            $ref: '#/components/schemas/app__models__views__CustomerPreferenceItem'
          type: array
          title: Preferences
      type: object
      required:
      - customer
      - activePreorders
      - recentOrders
      - preferences
      title: Customer360View
    CustomerDetail:
      properties:
        customerId:
          type: string
          title: Customerid
        customerCode:
          type: string
          title: Customercode
        fullName:
          type: string
          title: Fullname
        phone:
          type: string
          title: Phone
        status:
          $ref: '#/components/schemas/CustomerStatus'
        createdAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdat
        tags:
          items:
            type: string
          type: array
          title: Tags
        channelSource:
          anyOf:
          - type: string
          - type: 'null'
          title: Channelsource
        defaultAddress:
          anyOf:
          - type: string
          - type: 'null'
          title: Defaultaddress
        district:
          anyOf:
          - type: string
          - type: 'null'
          title: District
        province:
          anyOf:
          - type: string
          - type: 'null'
          title: Province
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        lastOrderAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastorderat
        preferences:
          items:
            $ref: '#/components/schemas/app__models__customers__CustomerPreferenceItem'
          type: array
          title: Preferences
        duplicateCandidates:
          items:
            $ref: '#/components/schemas/CustomerDuplicateCandidateSummary'
          type: array
          title: Duplicatecandidates
      type: object
      required:
      - customerId
      - customerCode
      - fullName
      - phone
      - status
      title: CustomerDetail
...
    ErrorResponse:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        correlationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlationid
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
      type: object
      required:
      - code
      - message
      title: ErrorResponse
...
    FarmSummaryBoardResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/FarmSummaryBoardItem'
          type: array
          title: Items
      type: object
      required:
      - items
      title: FarmSummaryBoardResponse
    FarmView:
      properties:
        plots:
          items:
            $ref: '#/components/schemas/PlotSummary'
          type: array
          title: Plots
        cropCycles:
          items:
            $ref: '#/components/schemas/CropCycleSummary'
          type: array
          title: Cropcycles
      type: object
      required:
      - plots
      - cropCycles
      title: FarmView
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Meta:
      properties:
        correlationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlationid
        causationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Causationid
        actorId:
          anyOf:
          - type: string
          - type: 'null'
          title: Actorid
        actorRole:
          anyOf:
          - type: string
          - type: 'null'
          title: Actorrole
        idempotencyKey:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotencykey
        externalRef:
          anyOf:
          - type: string
          - type: 'null'
          title: Externalref
        bypassRequested:
          type: boolean
          title: Bypassrequested
          default: false
        delegatedActorId:
          anyOf:
          - type: string
          - type: 'null'
          title: Delegatedactorid
        delegatedActorRole:
          anyOf:
          - type: string
          - type: 'null'
          title: Delegatedactorrole
      type: object
      title: Meta
...
    PendingFulfillmentListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PendingFulfillmentView'
          type: array
          title: Items
      type: object
      required:
      - items
      title: PendingFulfillmentListResponse
...
    ProjectContributionSummaryResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectContributionSummaryItem'
          type: array
          title: Items
      type: object
      required:
      - items
      title: ProjectContributionSummaryResponse
...
    ProjectPnlSummaryResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectPnlSummaryItem'
          type: array
          title: Items
      type: object
      required:
      - items
      title: ProjectPnlSummaryResponse
...
    ProjectOrderAllocationSummaryResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectOrderAllocationSummaryItem'
          type: array
          title: Items
      type: object
      required:
      - items
      title: ProjectOrderAllocationSummaryResponse
...
    ProjectContributionLedgerResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectContributionLedgerItem'
          type: array
          title: Items
      type: object
      required:
      - items
      title: ProjectContributionLedgerResponse
...
    ProjectImpactedActorSummaryResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectImpactedActorSummaryItem'
          type: array
          title: Items
      type: object
      required:
      - items
      title: ProjectImpactedActorSummaryResponse
...
    ProjectScopeDetail:
      properties:
        projectScopeId:
          type: string
          title: Projectscopeid
        organizationId:
          type: string
          title: Organizationid
        projectScopeCode:
          type: string
          title: Projectscopecode
        name:
          type: string
          title: Name
        projectScopeType:
          $ref: '#/components/schemas/ProjectScopeType'
        status:
          $ref: '#/components/schemas/ProjectScopeStatus'
        seasonYear:
          anyOf:
          - type: string
          - type: 'null'
          title: Seasonyear
        ownerActorId:
          anyOf:
          - type: string
          - type: 'null'
          title: Owneractorid
        createdAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdat
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        parentProjectScopeId:
          anyOf:
          - type: string
          - type: 'null'
          title: Parentprojectscopeid
        metadata:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Metadata
        updatedAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Updatedat
      type: object
      required:
      - projectScopeId
      - organizationId
      - projectScopeCode
      - name
      - projectScopeType
      - status
      title: ProjectScopeDetail
...
    ProjectScopeResponse:
      properties:
        data:
          $ref: '#/components/schemas/ProjectScopeDetail'
      type: object
      required:
      - data
      title: ProjectScopeResponse
    ProjectScopeStatus:
      type: string
      enum:
      - draft
      - active
      - paused
      - closed
      - archived
      title: ProjectScopeStatus
...
    ProjectScopeType:
      type: string
      enum:
      - value_stream
      - product_line
      - household_livelihood
      - experience
      - campaign
      - umbrella_program
      - shared_service
      title: ProjectScopeType
    SharedResourceDetail:
      properties:
        sharedResourceId:
          type: string
          title: Sharedresourceid
        organizationId:
          type: string
          title: Organizationid
        resourceCode:
          type: string
          title: Resourcecode
        name:
          type: string
          title: Name
        resourceType:
          $ref: '#/components/schemas/SharedResourceType'
        status:
          $ref: '#/components/schemas/SharedResourceStatus'
        capacityValue:
          anyOf:
          - type: number
          - type: 'null'
          title: Capacityvalue
        capacityUnit:
          anyOf:
          - type: string
          - type: 'null'
          title: Capacityunit
        createdAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdat
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        updatedAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Updatedat
      type: object
      required:
      - sharedResourceId
      - organizationId
      - resourceCode
      - name
      - resourceType
      - status
      title: SharedResourceDetail
    SharedResourceListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SharedResourceSummary'
          type: array
          title: Items
      type: object
      required:
      - items
      title: SharedResourceListResponse
    SharedResourceResponse:
      properties:
        data:
          $ref: '#/components/schemas/SharedResourceDetail'
      type: object
      required:
      - data
      title: SharedResourceResponse
    SharedResourceStatus:
      type: string
      enum:
      - draft
      - active
      - inactive
      - retired
      title: SharedResourceStatus
    SharedResourceSummary:
      properties:
        sharedResourceId:
          type: string
          title: Sharedresourceid
        organizationId:
          type: string
          title: Organizationid
        resourceCode:
          type: string
          title: Resourcecode
        name:
          type: string
          title: Name
        resourceType:
          $ref: '#/components/schemas/SharedResourceType'
        status:
          $ref: '#/components/schemas/SharedResourceStatus'
        capacityValue:
          anyOf:
          - type: number
          - type: 'null'
          title: Capacityvalue
        capacityUnit:
          anyOf:
          - type: string
          - type: 'null'
          title: Capacityunit
        createdAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdat
      type: object
      required:
      - sharedResourceId
      - organizationId
      - resourceCode
      - name
      - resourceType
      - status
      title: SharedResourceSummary
    SharedResourceType:
      type: string
      enum:
      - labor_pool
      - vehicle
      - drying_yard
      - warehouse
      - marketing_budget
      - content_asset
      - host_capacity
      - other
      title: SharedResourceType
    CreateSharedResourceRequest:
      properties:
        organizationId:
          type: string
          title: Organizationid
        name:
          type: string
          title: Name
        resourceType:
          $ref: '#/components/schemas/SharedResourceType'
        capacityValue:
          anyOf:
          - type: number
          - type: 'null'
          title: Capacityvalue
        capacityUnit:
          anyOf:
          - type: string
          - type: 'null'
          title: Capacityunit
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        meta:
          anyOf:
          - $ref: '#/components/schemas/Meta'
          - type: 'null'
      type: object
      required:
      - organizationId
      - name
      - resourceType
      title: CreateSharedResourceRequest
...
    UpdateProjectScopeRequest:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        seasonYear:
          anyOf:
          - type: string
          - type: 'null'
          title: Seasonyear
        ownerActorId:
          anyOf:
          - type: string
          - type: 'null'
          title: Owneractorid
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        parentProjectScopeId:
          anyOf:
          - type: string
          - type: 'null'
          title: Parentprojectscopeid
        metadata:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Metadata
        meta:
          anyOf:
          - $ref: '#/components/schemas/Meta'
          - type: 'null'
      type: object
      title: UpdateProjectScopeRequest
...
    app__models__views__CustomerPreferenceItem:
      properties:
        preferenceType:
          type: string
          title: Preferencetype
        preferenceValue:
          type: string
          title: Preferencevalue
        confidenceLevel:
          type: number
          title: Confidencelevel
      type: object
      required:
      - preferenceType
      - preferenceValue
      - confidenceLevel
      title: CustomerPreferenceItem